introduce recordset provider + skeleton for choice facet + facet picker
